Move the database to Cloudflare D1 and prepare Vercel deployment
- Port the schema from Postgres to SQLite
- Production reaches D1 over Cloudflare's HTTP API; development uses local.db via libsql
- Regenerate migrations; chunk seed inserts under D1's bound-value limit
- Framework preset for Vercel; README covers the new database setup
- Ignore api.txt and the local database file

+++ src/app/admin/page.tsx
@@ -15,7 +15,7 @@ export const dynamic = 'force-dynamic';
 
 async function countRows(table: unknown, where?: unknown) {
   const db = await getDynamicDb();
-  const q = db.select({ n: sql<number>`count(*)::int` }).from(table);
+  const q = db.select({ n: sql<number>`count(*)` }).from(table);
   const [row] = where ? await q.where(where) : await q;
   return Number(row?.n ?? 0);
 }
@@ -62,7 +62,7 @@ export default async function Overview({
 
   // Portfolio split by kind, for the part-to-whole bar.
   const kindRows = await db
-    .select({ kind: properties.kind, n: sql<number>`count(*)::int` })
+    .select({ kind: properties.kind, n: sql<number>`count(*)` })
     .from(properties)
     .groupBy(properties.kind);
